Create a simple Route 53 Lambda backup script
Creates a basic Python 3 script to be run as an AWS Lambda function or
cron job to back up Route 53 DNS records to an AWS S3 bucket. While
functional, this script is somewhat simplistic and crude, and lacks
meaningful documentation.

Going forward, the code should be refactored to make it
cleaner and more readable. Documentation and comments should also be
added and expanded. Making the code backwards-compatible with Python 2.7
should also be considered.

Future development should also include a Python or shell script to
upload this script and create the Lambda function, complete with IAM
role and environmental variables.

The backup csv files generated by this script can be restored using the
route53-transfer module. The code and documentation for this module,
including how to restore the Route 53 DNS records, can be found here:

https://github.com/RisingOak/route53-transfer