Remove redundant function/API calls
Eliminates redundant function/API calls to fetch zone records from the
write functions. This information is now read once and passed to the
different write functions to be parsed there, resulting in a shorter
overall function duration.

Write functions now return the name of the file they write to, allowing
that name to be called directly and reducing redundancy and the
possiblity of errors if the file name changes in the future.