Add error handling for ClientError exception
When creating S3 buckets, route53-lambda-backup now tests whether the
bucket exists before trying to create it and explicitly handles the
ClientError exception should it be raised.

If the bucket exists but returns a 403: Forbidden error, it is owned by
another user or otherwise unaccessible, and route53-lambda-backup exits
immediately and returns False. Otherwise, the bucket is created (if it
does not already exist) and route53-lambda-backup returns True.